WIP: use Sldr.Cleanup after tests
so that it isn't already initialised for later tests

There is a general issue with the tests now that a failure which doesn't
clean up sldr will cause lots of later tests to fail

+++ src/LexicalTools.Tests/MissingInfoTaskTests.cs
@@ -82,8 +82,15 @@ namespace WeSay.LexicalTools.Tests
 		[TearDown]
 		public void TearDown()
 		{
-			_lexEntryRepository.Dispose();
-			_tempFolder.Delete();
+			if (_lexEntryRepository != null)
+			{
+				_lexEntryRepository.Dispose();
+			}
+			if (_tempFolder != null)
+			{
+				_tempFolder.Delete();
+			}
+			WeSayProjectTestHelper.CleanupForTests();
 		}
 
 		[Test]
